new error appeared during push the MockTest
Hello
In the Sunday morning, I started solving the MockTest after I completed the full tutorials for learning Python and Pandas. 
I completed the MockTest and for more than 10 hours , I tried to fix the OSError:pytest in my computer but I could not be able. On Friday I pushed the Q0 and it looked working.
Now, I uploaded the files and tested. The codes are working excellent and the results are perfect as requesting.
It is so painful when working so much hard and spent studying around 16 hours per day and error like that destroying all that effort.
I hope you accept it and I hope the error will fix by the real exam.